requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
python-jose==3.3.0
bcrypt==4.0.1
requests==2.31.0
flask-limiter==3.5.0
openai==1.6.1
python-dotenv==1.0.0
google-api-python-client==2.108.0
